requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
astroid==3.3.10
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
dill==0.4.0
Faker==37.4.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.2
pylint==3.3.7
pytest==8.4.1
requests==2.32.4
tomlkit==0.13.3
tzdata==2025.2
urllib3==2.5.0
utils==1.0.2
